requirements.txt
streamlit
langchain
beautifulsoup4
requests
python-dotenv
langchain-groq 